sample-app/requirements.txt
flask==2.3.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-exporter-otlp==1.20.0
boto3==1.26.0
requests==2.31.0